+++ lib/clients/metaApi/metatraderAccount_client.py
@@ -98,8 +98,7 @@ class MetatraderAccountDto(TypedDict):
     metadata: Dict
     """Extra information which can be stored together with your account."""
     reliability: str
-    """Used to increase the reliability of the account. Allowed values are regular, high and maximum.
-    Default is regular."""
+    """Used to increase the reliability of the account. Allowed values are regular and high. Default is regular."""
 
 
 class NewMetatraderAccountDto(TypedDict):
@@ -136,8 +135,7 @@ class NewMetatraderAccountDto(TypedDict):
     metadata: Dict
     """Extra information which can be stored together with your account."""
     reliability: str
-    """Used to increase the reliability of the account. Allowed values are regular, high and maximum.
-    Default is regular."""
+    """Used to increase the reliability of the account. Allowed values are regular and high. Default is regular."""
 
 
 class MetatraderAccountUpdateDto(TypedDict):