+++ changelog.md
@@ -1,3 +1,9 @@
+12.3.0
+  - added credit account property
+  - added feature to unsubscribe from market data (remove symbol from market watch)
+  - removed maximum reliability value
+  - fixed synchronization throttling
+
 12.2.0
   - added retryOpts option to configure retries of certain REST/RPC requests upon failure
   - improved socket connection reliability